Push

+++ app.js
@@ -1,5 +1,8 @@
 (function () {
 
+
+	'use strict';
+
 	if (window.applicationCache) {
 		applicationCache.addEventListener('updateready', function() {
 			if (confirm('An update is available. Reload now?')) {
@@ -9,51 +12,54 @@
 		});
 	}
 
+	Offline.options = {
+		checks: {xhr: {url: 'https://cdnjs.cloudflare.com/ajax/libs/offline-js/0.7.14/offline.min.js'}}
+	};
+
+	// setInterval(Offline.check, 500);
+
 })();
 
-
-'use strict';
 
 var app = angular.module('creepy-octo-fibula', ['ngResource', 'ngSanitize', 'ui.router', 'firebase']);
 
 app.constant('FIREBASE_URL', 'https://creepy-octo-fibula.firebaseio.com/');
 
 app.run(['$rootScope', '$log', function ($rootScope, $log) {
+
+
 	$log.log(Offline.state);
 
-	localforage.setDriver(localforage.LOCALSTORAGE);
 
-	localforage.getItem('sync').then(function(value) {
-		if (value === null) {
-			$log.error("Sync does not exists");
-			localforage.setItem('sync', false, function (err, value) {
-				console.log(value);
-			});
-		} else {
-			localforage.getItem('sync').then(function(value) {
-				if (value === true) {
-
-				} else {
-
-				}
-			});
+	if (localStorage.getItem('sync') === null) {
+		$log.error("Sync does not exists");
+		localStorage.setItem('sync', false);
+	} else {
+		var sync = localStorage.getItem('sync');
+		if (sync === "true" && Offline.state === "up") {
+			syncmessenger();
 		}
-	});
+	}
 
-	localforage.getItem('merge').then(function(value) {
-		if (value === null) {
-			var data = [];
-			localforage.setItem('merge', JSON.stringify(data), function (err, value) {
-				console.log(value);
-			});
-		} else {
-			localforage.getItem('sync').then(function (err, value) {
-				if(value) {
-					syncmessenger();
-				}
-			});
-		}
-	});
+	if (localStorage.getItem('merge') === null) {
+		var data = [];
+		localStorage.setItem('merge', JSON.stringify(data));
+	}
+
+	if (localStorage.getItem('newdata') === null) {
+		var data = [];
+		localStorage.setItem('newdata', JSON.stringify(data));
+	}
+
+	if (localStorage.getItem('fuel') === null) {
+		var data = [];
+		localStorage.setItem('fuel', JSON.stringify(data));
+	}
+
+	if (localStorage.getItem('delete') === null) {
+		var data = [];
+		localStorage.setItem('delete', JSON.stringify(data));
+	}
 
 
 }]);
@@ -85,15 +91,30 @@ app.controller('ListController', ['$scope', '$log', 'ListService', function ($sc
 		$scope.messages = ListService.getAssets();
 
 		ListService.getAssets().$loaded().then(function (x) {
-			localforage.setItem('merge', JSON.stringify(x), function (err, value) {
-				console.log("Synced Pull Data ", value);
-			});
+			localStorage.setItem('merge', JSON.stringify(x));
 		});
+
+		$scope.newdatas = ListService.getLocalData();
 	} else {
-		localforage.getItem('merge').then(function (err, value) {
-			$scope.messages = JSON.parse(value);
-		});
+		$scope.messages = JSON.parse(localStorage.getItem('merge'));
+		$scope.newdatas = ListService.getLocalData();
 	}
+
+	$scope.$watch(function() {
+		return ListService.getMergeData();
+	}, function(value) {
+		if(value) {
+			$scope.messages = ListService.getMergeData();
+		}
+	}, true);
+
+	$scope.$watch(function() {
+		return ListService.getLocalData();
+	}, function(value) {
+		if(value) {
+			$scope.newdatas = ListService.getLocalData();
+		}
+	}, true);
 
 
 	$scope.imagePath = function (assetid) {
@@ -105,8 +126,31 @@ app.controller('ListController', ['$scope', '$log', 'ListService', function ($sc
 	};
 
 	$scope.deleteAsset = function (id) {
-		ListService.deleteAsset(id);
+		if (Offline.state === "up") {
+			console.log("Delete Data - Online");
+			localStorage.setItem('sync', false);
+			ListService.deleteAsset(id);
+		} else {
+			console.log("Delete Data - Offline");
+			localStorage.setItem('sync', true);
+			var deldata = JSON.parse(localStorage.getItem('delete'));
+			deldata.push(id);
+			localStorage.setItem('delete', JSON.stringify(deldata));
+
+			var merge = JSON.parse(localStorage.getItem('merge'));
+			var newmerge = [];
+
+			angular.forEach(merge, function(message, key) {
+				if (message.id !== id) {
+					 newmerge.push(message);
+				}
+			});
+
+			localStorage.setItem('merge', JSON.stringify(newmerge));
+
+		}
 	};
+
 
 }]);
 
@@ -114,6 +158,13 @@ app.factory('ListService', ['$firebaseArray', 'FIREBASE_URL', function ($firebas
 
 	var exports = {};
 	var reference = new Firebase(FIREBASE_URL);
+	reference.on("value", function(snap) {
+		if (snap.val() === true) {
+			alert("connected");
+		} else {
+			alert("not connected");
+		}
+	});
 
 	exports.getAssets = function () {
 		var messages = $firebaseArray(reference);
@@ -138,10 +189,9 @@ app.factory('ListService', ['$firebaseArray', 'FIREBASE_URL', function ($firebas
 
 	exports.updateFuelValue = function (id, value) {
 
-
-		localforage.setItem('merge', function (err, value) {
-			console.log(value);
-		});
+		var data = JSON.parse(localStorage.getItem('fuel'));
+		data.push({"id": id, "fuel": value});
+		localStorage.setItem('fuel', JSON.stringify(data));
 
 		if (Offline.state === "up") {
 			var messages = $firebaseArray(reference);
@@ -157,34 +207,99 @@ app.factory('ListService', ['$firebaseArray', 'FIREBASE_URL', function ($firebas
 						});
 					}
 				});
+				localStorage.setItem('merge', JSON.stringify(messages));
+				data = [];
+				localStorage.setItem('fuel', JSON.stringify(data));
 			});
 
-			localforage.setItem('sync', false, function (err, value) {
-				console.log(value);
-			});
+			localStorage.setItem('sync', false);
 		} else {
-			localforage.setItem('sync', true, function (err, value) {
-				console.log(value);
+
+			var merge = JSON.parse(localStorage.getItem('merge'));
+
+			angular.forEach(merge, function(message, key) {
+				if (message.id === id) {
+					message.fuel = value;
+					console.log(key, message);
+				}
 			});
 
-			localforage.setItem('merge', function (err, value) {
-				console.log(value);
-			});
+			localStorage.setItem('merge', JSON.stringify(merge));
+			localStorage.setItem('sync', true);
 		}
 	};
 
 	exports.deleteAsset = function (id) {
-		var messages = $firebaseArray(reference);
 
-		messages.$loaded().then(function(x) {
-			angular.forEach(x, function(message, key) {
-				if (message.id === id) {
-					messages.$remove(key).then(function (argument) {
-						console.log("Delete");
-					});
-				}
+			var messages = $firebaseArray(reference);
+
+			messages.$loaded().then(function(x) {
+				angular.forEach(x, function(message, key) {
+					if (message.id === id) {
+						messages.$remove(key).then(function (argument) {
+							console.log("Delete");
+						});
+					}
+				});
 			});
+	};
+
+	exports.updateAll = function () {
+
+		var newdata = JSON.parse(localStorage.getItem('newdata'));
+
+		angular.forEach(newdata, function (message, key) {
+			exports.addAsset(message);
 		});
+
+
+		var empty = [];
+		localStorage.setItem('newdata', JSON.stringify(empty));
+
+		var deldata = JSON.parse(localStorage.getItem('delete'));
+
+		angular.forEach(deldata, function (message, key) {
+			exports.deleteAsset(message);
+		});
+
+		localStorage.setItem('delete', JSON.stringify(empty));
+
+		exports.getAssets().$loaded().then(function (x) {
+			localStorage.setItem('merge', JSON.stringify(x));
+		});
+
+		var fueldata = JSON.parse(localStorage.getItem('fuel'));
+
+		if (fueldata.length > 0) {
+			var messages = $firebaseArray(reference);
+
+			messages.$loaded().then(function(x) {
+				angular.forEach(fueldata, function (v, i) {
+					angular.forEach(x, function(message, key) {
+						if (message.id === v.id) {
+							message.fuel = v.fuel;
+							console.log(key, message);
+							messages.$save(key).then(function (argument) {
+								console.log("Saved");
+								// Saved Message
+							});
+						}
+					});
+				});
+				localStorage.setItem('merge', JSON.stringify(messages));
+				var data = [];
+				localStorage.setItem('fuel', JSON.stringify(data));
+			});
+		}
+
+	};
+
+	exports.getLocalData = function () {
+		return JSON.parse(localStorage.getItem('newdata'));
+	};
+
+	exports.getMergeData = function () {
+		return JSON.parse(localStorage.getItem('merge'));
 	};
 
 	return exports;
@@ -194,9 +309,29 @@ app.factory('ListService', ['$firebaseArray', 'FIREBASE_URL', function ($firebas
 app.controller('AddController', ['$scope', '$log', 'ListService', function ($scope, $log, ListService) {
 	console.log("Add");
 
+	function guidGenerator() {
+		var S4 = function() {
+			return (((1+Math.random())*0x10000)|0).toString(16).substring(1);
+		};
+		return (S4()+S4()+"-"+S4()+"-"+S4()+"-"+S4()+"-"+S4()+S4()+S4());
+	};
+
 	$scope.addAsset = function () {
+
 		$scope.formdata.assetid = Math.floor((Math.random()*8)+1);
-		ListService.addAsset($scope.formdata);
+
+		if (Offline.state === "up") {
+			console.log("New Data - Online");
+			localStorage.setItem('sync', false);
+			ListService.addAsset($scope.formdata);
+		} else {
+			console.log("New Data - Offline");
+			$scope.formdata.id = guidGenerator();
+			localStorage.setItem('sync', true);
+			var newdata = JSON.parse(localStorage.getItem('newdata'));
+			newdata.push($scope.formdata);
+			localStorage.setItem('newdata', JSON.stringify(newdata));
+		}
 	};
 
 }]);
@@ -204,21 +339,20 @@ app.controller('AddController', ['$scope', '$log', 'ListService', function ($sco
 Offline.on('confirmed-up', function () {
 	console.info("Confirmed Up");
 
-	localforage.getItem('sync').then(function(value) {
-		if (value) {
-			syncmessenger();
-			console.info("Need to sync");
-		} else {
-		}
-	});
-
+	var sync = localStorage.getItem('sync');
+	if (sync === "true") {
+		syncmessenger();
+		console.info("Need to sync");
+	}
 });
+
 
 Offline.on('confirmed-down', function () {
 	console.info("Confirmed Down");
 });
 
 function syncmessenger() {
+	Messenger().hideAll();
 	var msg = Messenger().post({
 		message: "Do you like to sync?",
 		hideAfter: 10000,
@@ -227,9 +361,9 @@ function syncmessenger() {
 				label: 'Sync Now',
 				action: function() {
 					console.log("Syncing Local Data");
-					var senddata = JSON.parse(localStorage.getItem('local-data'));
-					senddata.sync = false;
-					updateServerData(senddata);
+					angular.injector(['creepy-octo-fibula']).get('ListService').updateAll();
+
+					localStorage.setItem('sync', false);
 					Messenger().hideAll();
 				}
 			},
